delete: Root.tsx 삭제하고 App.tsx로 합치고, shared파일에서 뺌
