--- FE/airbnb/src/client/Root.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import { BrowserRouter } from "react-router-dom";
-import App from "shared/App";
-
-function Root() {
-  return (
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
-  );
-}
-
-export default Root;
